Change Project Structure
Change Project to use netstandard2.0
- This can be changed to net net472 later if needed.
Updated README to reflect new setup instructions
