+++ StormFlake.Common.props
@@ -1,0 +1,41 @@
+<Project>
+    <PropertyGroup>
+        <AssemblyName>StormFlake</AssemblyName>
+        <AssemblyVersion>1.0.0</AssemblyVersion>
+        <ReleaseVersion>1.0.0</ReleaseVersion>
+        <TargetFramework>netstandard2.0</TargetFramework>
+        <LangVersion>latest</LangVersion>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+        <DebugType>full</DebugType>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <BepInExPluginGuid Condition="'$(BepInExPluginGuid )' == ''">$(Guid)</BepInExPluginGuid>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <!-- First one is "Forced Access" warning, second one is "Missing XML" comment -->
+        <NoWarn>$(NoWarn);Publicizer001;CS1591</NoWarn>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <OutputPath><!-- Path To Your Bepinex Install !-->\plugins\StormFlake\</OutputPath>
+        <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <DocumentationFile>$(ProjectDir)bin/$(Configuration)/$(MSBuildProjectName)/$(MSBuildProjectName).xml</DocumentationFile>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="BepInEx.Analyzers" Version="1.0.8" PrivateAssets="all" />
+        <PackageReference Include="BepInEx.Core" Version="5.4.21" />
+        <PackageReference Include="BepInEx.PluginInfoProps" Version="2.1.0" />
+        <PackageReference Include="UnityEngine.Modules" Version="2021.3.16" IncludeAssets="compile" />
+    </ItemGroup>
+
+    <ItemGroup Condition="'$(TargetFramework.TrimEnd(`0123456789`))' == 'net'">
+        <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.3" PrivateAssets="all" />
+    </ItemGroup>
+
+</Project>
