--- StormFlake/MonoBehaviours/StormFlakeComponent.cs
@@ -1,46 +1,0 @@
-﻿using UnityEngine;
-
-namespace StormFlake.MonoBehaviours
-{
-    // TODO Review this file and update to your own requirements, or remove it altogether if not required
-    /// <summary>
-    /// Template MonoBehaviour class. Use this to add new functionality and behaviours to
-    /// the game.
-    /// </summary>
-    internal class StormFlakeComponent : MonoBehaviour
-    {
-        /// <summary>
-        /// Unity Awake method.
-        /// </summary>
-        public void Awake()
-        {
-
-        }
-
-        /// <summary>
-        /// Unity Start method
-        /// </summary>
-        public void Start()
-        {
-
-        }
-
-        /// <summary>
-        /// Unity Awake method. Runs every frame so remove this if not required.
-        /// Runs frequently, so remove if not required.
-        /// </summary>
-        public void Update()
-        {
-
-        }
-
-        /// <summary>
-        /// Unity Physics Update (LateUpdate) method.
-        /// Runs frequently, so remove if not required.
-        /// </summary>
-        public void LateUpdate()
-        {
-
-        }
-    }
-}
